backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pymongo==4.5.0
Flask-PyMongo==2.3.0
bcrypt==4.0.1
python-dotenv==1.0.0
azure-storage-blob==12.19.0
hypothesis==6.88.1
pytest==7.4.2
pytest-flask==1.3.0